prbt_support: drop all test code
There is no compiled code in the package, so drop all compiler references.

The acceptance_test requires real hardware and doesn't make sense for MoveIt.

The compiled URDF test is inappropriate as well, because the repository should
only provide descriptions/configurations for testing in main moveit.
The actual test was already disabled because it relies on moveit datatypes,
but the package.xml/find_package entries remained.